fixed bgez

+++ src/sim.c
@@ -366,7 +366,7 @@ void execute()
         // branch on greater than Or equal to zero
         // if ($s >= 0) pc += i << 2
         case OP_BGEZ:
-            if(CURRENT_STATE.REGS[dcd_rs] >= 0){
+            if((int32_t)CURRENT_STATE.REGS[dcd_rs] >= 0){
                 NEXT_STATE.PC = CURRENT_STATE.PC + (dcd_se_imm << 2);
             }
             break;
